admin: Move user list rows into own file.
