+++ src/admin/components/UsersList/index.js
@@ -6,40 +6,12 @@ import {
   TableBody,
   TableHeader,
   TableHeaderColumn,
-  TableRow,
-  TableCell
+  TableRow
 } from '../../../components/Table';
-import Avatar from '../../../components/Avatar';
-import Username from '../../../components/Username/WithCard';
-import formatJoinDate from '../../../utils/formatJoinDate';
+import UserRow from './Row';
 
 const avatarStyle = {
   width: 24
-};
-
-const UserRow = ({
-  user
-}) => (
-  <TableRow>
-    <TableCell style={avatarStyle}>
-      <Avatar user={user} />
-    </TableCell>
-    <TableCell>
-      <Username user={user} />
-    </TableCell>
-    <TableCell>
-      {formatJoinDate(user.createdAt)}
-    </TableCell>
-    <TableCell>Email</TableCell>
-    <TableCell>
-      {user.roles.join(', ')}
-    </TableCell>
-    <TableCell />
-  </TableRow>
-);
-
-UserRow.propTypes = {
-  user: PropTypes.object.isRequired
 };
 
 const enhance = translate();
